photo_tools/requirements.txt
piexif
Pillow
pyheif

